# Rework unit tests for SEDS UI