--- a/services/ui-src/src/components/AddUser/AddUser.test.jsx
+++ b/services/ui-src/src/components/AddUser/AddUser.test.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { mount } from "enzyme";
+import { render, screen } from "@testing-library/react";
 import { Provider } from "react-redux";
 import configureMockStore from "redux-mock-store";
 
@@ -15,42 +15,42 @@ const mockStore = configureMockStore([]);
 describe("Test AddUser.jsx", () => {
   beforeEach(() => {
     store = mockStore(stateListMock);
-    wrapper = mount(
+    wrapper = render(
       <Provider store={store}>
         <BrowserRouter>
           <AddUser />
         </BrowserRouter>
       </Provider>
-    );
+    ).container;
   });
 
   test("Check the main div, with classname AddUser, exists", () => {
-    expect(wrapper.find(".page-header").length).toBe(1);
+    expect(wrapper.querySelector(".page-header")).toBeInTheDocument();
   });
 
   test("Check the Main component exists", () => {
-    expect(wrapper.find(".addUserMain").length).toBe(1);
+    expect(wrapper.querySelector(".addUserMain")).toBeInTheDocument();
   });
 
   test("Check that the User List link is available", () => {
-    expect(wrapper.find(".userList").length).toBe(3);
+    expect(wrapper.querySelector(".userList")).toBeInTheDocument();
   });
 
   test("Check the EUA ID input component renders", () => {
-    expect(wrapper.find(".euaID").length).toBe(1);
+    expect(wrapper.querySelector(".eua-ID")).toBeInTheDocument();
   });
 
   test("Check the Role Selector drop-down renders", () => {
-    expect(wrapper.find(".selectRole").length).toBe(1);
+    expect(screen.getByPlaceholderText("Select a Role")).toBeInTheDocument();
   });
 
   /* Will need to come back to this test as this drop-down is not shown until a Role is selected
    test("Check the State Selector drop-down renders", () => {
-        expect(wrapper.find(".selectState").length).toBe(1);
+        expect(screen.getByPlaceholderText("Select a State")).toBeInTheDocument();
     });
     */
 
   test("Check the Add User button renders", () => {
-    expect(wrapper.find(".createUser").length).toBe(2);
+    expect(screen.getByText("Add User", { selector: "button"})).toBeInTheDocument();
   });
 });