--- a/services/ui-src/src/setupTests.js
+++ b/services/ui-src/src/setupTests.js
@@ -2,7 +2,7 @@
 // allows you to do things like:
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
-import "@testing-library/jest-dom/extend-expect";
+import "@testing-library/jest-dom";
 
 jest.mock("aws-amplify", () => ({
   Auth: {
@@ -38,3 +38,7 @@ jest.mock("./utility-functions/constants", () => ({
   MODE: "production",
   BASE_URL: "mdctcartsdev.cms.gov"
 }));
+
+HTMLCanvasElement.prototype.getContext = () => {
+  /* Nothing; we don't unit test our jsPDF integration */
+}