### Description
SEDS received a security warning concerning the package it used for unit test rendering, `enzyme`. We took this as an opportunity to switch to `testing-library`, to bring SEDS in line with the rest of the MDCT app suite.

This was not a trivial update. It involved changes to 172 unit tests. I took this as an opportunity to improve the quality of the test as well. The result is better test coverage, using fewer tests, running faster, verifying more of the app's behavior.

"Verifying more behavior" is difficult to quantify; take `EditUser` as an example. This is an admin page that shows user details (name, email, etc) in a set of editable inputs, with a save button below. Previously this component had 3 tests, each of which verified that a single element existed on the page, and only tested the page's empty state (when the user is not found). Now we have 5 tests, which verify:
* The same "element should exist" checks as the old tests
* That, when a user is found, all of their details are rendered correctly
   - Additionally, I discovered a labeling inconsistency, which I fixed
* That the page renders slightly differently when the user is an admin
   - (The user state dropdown becomes a multiselect, with the correct options displayed)
* That, when the details are edited and the save button is clicked, the correct API call is made
* The same "user not found" page state as the old tests
   - Additionally, this led to my discovering a small bug that would cause console errors for this page state, which I fixed

The test run time is down from about 11 seconds from to 9, on my machine. The test count is down from the 170s to the 130s. And the line coverage is up from 72% to 82%. Here's a table breaking it down file-by-file:

File                         | Before |  Delta |  After |
-----------------------------|--------|--------|--------|
All files                    |  72.05 |  10.82 |  82.87 |
  AddUser.jsx                |  43.33 |  30    |  73.33 |
  App.jsx                    |  38.09 |  61.91 | 100    |
  AuthenticatedRoute.jsx     | 100    |   0    | 100    |
  CertificationTab.jsx       |  64.44 |   0    |  64.44 |
  EditUser.jsx               |  39.02 |  28.05 |  67.07 |
  Footer.jsx                 | 100    |   0    | 100    |
  FormFooter.jsx             |  81.25 |   0    |  81.25 |
  FormHeader.jsx             |  80    |   0    |  80    |
  FormLoadError.jsx          | 100    |   0    | 100    |
  FormPage.jsx               |  45.71 |  45.71 |  91.42 |
  FormTemplates.jsx          |  30.43 |  47.83 |  78.26 |
  GREGridWithTotals.jsx      |  89.78 |   0    |  89.78 |
  GenerateForms.jsx          |  66.66 |  33.34 | 100    |
  GenerateTotals.jsx         |  66.66 |  33.34 | 100    |
  GridWithTotals.jsx         |  91.35 | -25.92 |  65.43 |
  Header.jsx                 |  73.33 |   0    |  73.33 |
  Home.jsx                   |  68.75 |  31.25 | 100    |
  HomeAdmin.jsx              |  81.08 |   0    |  81.08 |
  HomeState.jsx              |  90.47 |   9.53 | 100    |
  LoaderButton.jsx           | 100    |   0    | 100    |
  Login.jsx                  |  61.53 |   3.85 |  65.38 |
  NotApplicable.jsx          |  51.61 |   6.45 |  58.06 |
  NotFound.jsx               | 100    |   0    | 100    |
  Preloader.jsx              | 100    |   0    | 100    |
  PrintPDF.jsx               |  70.73 |  19.51 |  90.24 |
  Profile.jsx                |  40    |  26.61 |  66.66 |
  Quarterly.jsx              |  92.85 |   0    |  92.85 |
  Question.jsx               |  80    |   8    |  88    |
  Routes.jsx                 |   0    | 100    | 100    |
  Signup.jsx                 |  29.62 |  55.56 |  85.18 |
  StateSelector.jsx          |  48.38 |  41.94 |  90.32 |
  StatusButton.jsx           | 100    |   0    | 100    |
  SummaryNotes.jsx           |  83.33 |   4.17 |   87.5 |
  SummaryTab.jsx             | 100    |   0    | 100    |
  SynthesizedGrid.jsx        | 100    |   0    | 100    |
  SynthesizedGridSummary.jsx |  97.43 |   0    |  97.43 |
  TabContainer.jsx           |  97.22 |   2.78 | 100    |
  UnauthenticatedRoute.jsx   |  72.72 |  18.18 |  90.9  |
  Unauthorized.jsx           | 100    |   0    | 100    |
  Users.jsx                  |  56    |  44    | 100    |


### Related ticket(s)
CMDCT-3738

---
### How to test
```bash
cd services/ui-src
yarn
yarn test
```

### Notes
Bye, Enzyme.

---
### Pre-review checklist
<!-- Complete the following steps before opening for review -->
- [x] I have added [thorough](https://shorturl.at/aejkF) tests, if necessary
- [x] I have added [thorough](https://shorturl.at/aejkF) tests, if necessary
- [x] I have added [thorough](https://shorturl.at/aejkF) tests, if necessary
- [x] I have added [thorough](https://shorturl.at/aejkF) tests, if necessary
- ~[ ] I have updated relevant documentation, if necessary~
- [x] I have performed a self-review of my code
- [ ] I have manually tested this PR in the deployed cloud environment

---
### Pre-merge checklist
<!-- Complete the following steps before merging -->

#### Review
- [ ] Design: This work has been reviewed and approved by design, if necessary
- [ ] Product: This work has been reviewed and approved by product owner, if necessary

#### Security
_If either of the following are true, notify the team's ISSO (Information System Security Officer)._

- [ ] These changes are significant enough to require an update to the SIA.
- [ ] These changes are significant enough to require a penetration test.
---

<!-- If deploying to val or prod, click 'Preview' and select template -->
_convert to a different template: [test → val](?expand=1&template=test-to-val-deployment.md)_ | _[val → prod](?expand=1&template=val-to-prod-deployment.md)_